fixing missing button popovers
When sections of DOM are removed, popover objects linked to
the elements stop working. So instead of adding / removing
chunks of DOM we now hide them. It means that form fields from
hidden parts are also sent, but it's ok for now. We can remove
DOM with form parts later.